--- a/scripts/performance/.gitignore
+++ b/scripts/performance/.gitignore
@@ -1,0 +1,2 @@
+profiles
+workspace