… packages/core/README.md