@sdirix 2021-07-23
Please check whether you can add
```
// @ts-check
```
to the script