@sgraband 2021-07-23
I updated my changes with your feedback. Regarding the error you mentioned: This should not happen anymore. (I added a npm preference, that prepends the node path)

Additionally i enabled the ` @ts-check`. There was however a small problem with the puppeteer selectors, so i ignored those lines.