@sdirix 2021-07-23
The `yarn.lock` changes can be removed