@sgraband 2021-07-26
Removed the `yarn.lock` changes and updated the `package.json`.