@sgraband 2021-07-26
Rebased and added the changelog entry (PR number missing, as we do not have it yet)