@sdirix 2022-11-16
Merged upstream with https://github.com/eclipse-theia/theia/pull/9777